review: split in files, improve docs
